selecity update
